// ui/src/generated/RConfiguration.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RProbe } from "./RProbe";
import type { RSelection } from "./RSelection";

export type RConfiguration = { source: string, current: RProbe | null, candidates: Array<RSelection>, error: string | null, };